requirements.txt
streamlit>=1.34
pandas>=2.0
numpy>=1.25
python-dateutil>=2.9
